Migrated code to ReceiptHelper.
